4/8 좋아요 유저정보 저장기능 추가

+++ src/main/java/com/where/service/PlaceService.java
@@ -45,7 +45,7 @@ public class PlaceService {
         Meeting target = findService.valid(meetingRepository.findById(meetingId),EntityType.Meeting);
         User targetUser = findService.valid(userRepository.findById(userId),EntityType.User);
         if(!userMeetingMappingRepository.existsByUserAndMeeting(targetUser, target)) {
-            throw new IllegalArgumentException("해당 유저는 해당 모임의 구성원이 아닙니다!");
+            throw new IllegalArgumentException("해당 유저는 이 모임의 구성원이 아닙니다!");
         }
         List<GetPlaceResponse> places = new ArrayList<>();
         //해당 Meeting에 Place가 있는지 확인
@@ -78,7 +78,7 @@ public class PlaceService {
         User targetUser = findService.valid(userRepository.findById(userId), EntityType.User);
         log.warn(targetUser.toString());
         if(!userMeetingMappingRepository.existsByUserAndMeeting(targetUser, targetMeeting)) {
-            throw new IllegalArgumentException("해당 유저는 해당 모임에 속해있지 않습니다!");
+            throw new IllegalArgumentException("해당 유저는 이 모임의 구성원이 아닙니다!");
         }
         //카카오맵 api 이용해 해당 장소 정보 검색
         PlaceAddressDto placeAddressDto = apiService.getKakaoMapPlaceId(name, address);
@@ -193,6 +193,9 @@ public class PlaceService {
 
         //id 유효성 판단
         Place target = findService.valid(placeRepository.findById(placeId), EntityType.Place);
+        if(!userMeetingMappingRepository.existsByUserAndMeeting(findService.valid(userRepository.findById(userId),EntityType.User),target.getMeeting())) {
+            throw new IllegalArgumentException("해당 유저는 이 모임의 구성원이 아닙니다!");
+        }
         //장소공유한 유저정보 삭제
         userPlaceMappingRepository.deleteAllByPlace(target);
         //해당 장소에 저장되어있는 코멘트 삭제
@@ -215,6 +218,9 @@ public class PlaceService {
 
         //id 유효성 판단
         Place target = findService.valid(placeRepository.findById(placeId), EntityType.Place);
+        if(!userMeetingMappingRepository.existsByUserAndMeeting(findService.valid(userRepository.findById(userId),EntityType.User),target.getMeeting())) {
+            throw new IllegalArgumentException("해당 유저는 이 모임의 구성원이 아닙니다!");
+        }
         Place saved = placeRepository.save(Place.builder()
                     .id(target.getId())
                     .meeting(target.getMeeting())
@@ -249,7 +255,12 @@ public class PlaceService {
 
         //id 유효성 검증
         Place target = findService.valid(placeRepository.findById(placeId),EntityType.Place);
-        findService.valid(userRepository.findById(userId), EntityType.User);
+        User targetUser = findService.valid(userRepository.findById(userId), EntityType.User);
+
+        //구성원정보 확인
+        if(!userMeetingMappingRepository.existsByUserAndMeeting(targetUser,target.getMeeting())){
+            throw new IllegalArgumentException("해당유저는 이 모임의 구성원이 아닙니다!");
+        }
         boolean like = target.getLikes().contains(userId);
         //likes에 해당 유저가 있는지 확인후 토글
         if(like){
@@ -292,6 +303,10 @@ public class PlaceService {
         //각종 id 유효성 검사
         Place targetPlace = findService.valid(placeRepository.findById(placeId), EntityType.Place);
         User targetUser = findService.valid(userRepository.findById(userId), EntityType.User);
+        if(!userMeetingMappingRepository.existsByUserAndMeeting(targetUser,targetPlace.getMeeting())){
+            throw new IllegalArgumentException("해당 유저는 이 모임의 구성원이 아닙니다!");
+        }
+
         //해장 장소를 공유한 사용자가 맞는지 확인
         if(userPlaceMappingRepository.existsByUserAndPlace(targetUser,targetPlace)) {
             if(commentRepository.existsByUser(targetUser)){
@@ -333,6 +348,9 @@ public class PlaceService {
 
         //id 유효성 확인
         Comment target = findService.valid(commentRepository.findById(commentId), EntityType.Comment);
+        if(!userMeetingMappingRepository.existsByUserAndMeeting(findService.valid(userRepository.findById(userId),EntityType.User),target.getPlace().getMeeting())) {
+            throw new IllegalArgumentException("해당 유저는 이 모임의 구성원이 아닙니다!");
+        }
         //해당코멘트를 단 유저인지 확인
         if(target.getUser().getId().equals(userId)) {
             Comment saved = commentRepository.save(Comment.builder()
@@ -369,6 +387,10 @@ public class PlaceService {
 
         //id 유효성 검사
         Comment target = findService.valid(commentRepository.findById(commentId), EntityType.Comment);
+        if(!userMeetingMappingRepository.existsByUserAndMeeting(target.getUser(),target.getPlace().getMeeting())) {
+            throw new IllegalArgumentException("해당 유저는 이 모임의 구성원이 아닙니다!");
+        }
+
         if(target.getUser().getId().equals(userId)) {
             commentRepository.delete(target);
             //모임 구성원에게 데이터 전송
